refactor: move get_manga_preferred_scanlator to separate file
