+++ backend/shared/src/usecases/get_manga_preferred_scanlator.rs
@@ -1,0 +1,13 @@
+use anyhow::Result;
+use crate::{
+    database::Database,
+    model::{MangaId},
+};
+
+pub async fn get_manga_preferred_scanlator(
+    db: &Database, 
+    manga_id: &MangaId
+) -> Result<Option<String>> {
+    let state = db.find_manga_state(manga_id).await;
+    Ok(state.and_then(|s| s.preferred_scanlator))
+}
