+++ backend/shared/src/usecases/mod.rs
@@ -18,6 +18,7 @@ pub mod set_source_stored_settings;
 pub mod uninstall_source;
 pub mod update_settings;
 pub mod set_manga_preferred_scanlator;
+pub mod get_manga_preferred_scanlator;
 
 pub use add_manga_to_library::add_manga_to_library;
 pub use check_update::check_update;
@@ -38,7 +39,5 @@ pub use search_mangas::search_mangas;
 pub use set_source_stored_settings::set_source_stored_settings;
 pub use uninstall_source::uninstall_source;
 pub use update_settings::update_settings;
-pub use set_manga_preferred_scanlator::{
-    get_manga_preferred_scanlator,
-    set_manga_preferred_scanlator,
-};
+pub use set_manga_preferred_scanlator::set_manga_preferred_scanlator;
+pub use get_manga_preferred_scanlator::get_manga_preferred_scanlator;
