+++ backend/shared/src/usecases/set_manga_preferred_scanlator.rs
@@ -20,11 +20,3 @@ pub async fn set_manga_preferred_scanlator(
     
     Ok(())
 }
-
-pub async fn get_manga_preferred_scanlator(
-    db: &Database, 
-    manga_id: &MangaId
-) -> Result<Option<String>> {
-    let state = db.find_manga_state(manga_id).await;
-    Ok(state.and_then(|s| s.preferred_scanlator))
-}
